+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2021.07.19
+
+- Changed: License key is now provided by value instead of a file.
+- Changed: Activation, download CA public cert, and download jar tasks are standard tasks
+- Added: Playbook for Molecule verification
+
 ## 2021.07.12
 
 - Changed: Identify and download the latest Burp Suite installer version